Refactor list item
