+++ src/components/ListItems/ListItems.stories.js
@@ -1,21 +1,32 @@
 
 import React, { useState } from 'react';
-import { View, Alert } from 'react-native';
+import { View, Alert, StyleSheet, Text } from 'react-native';
 
 import { storiesOf } from '@storybook/react-native';
 
 import ListItems from './ListItems';
 
-const functionToOpenNewPage = () => {
-    return alert("Topic Detail")
-}
+const styles = StyleSheet.create({
+    topic: {
+        marginLeft: 18,
+        fontFamily: 'Inter',
+        fontStyle: 'normal',
+        fontWeight: 'normal',
+        fontSize: 16,
+        lineHeight: 19,
+        color: '#111827'
 
+    }
+})
 
 
 storiesOf('ListItem', module)
     .addDecorator(story => <View>{story()}</View>)
-    .add('List', () => <ListItems label={'List'} display={'flex'} />)
-    .add('ListLinking', () => <ListItems label={'ListLinking'} display={'flex'} onPress={functionToOpenNewPage} />);
+    .add('List of topic', () => <ListItems >
+        <Text style={styles.topic}> All </Text>
+    </ListItems>)
+    .add('List of seen phrases', () => <ListItems ><View><Text style={styles.topic}>Words and phrases</Text></View></ListItems>)
+
